+++ controllers/routers/rules/account.js
@@ -1,19 +1,20 @@
-module.exports = authentication = {};
+module.exports = exports = { account: {} };
 
 
 
 // ==============================
-//  FUNCTIONS
+//  DEPENDENCIES
 // ==============================
 const { _md } = require('../../middleware/plugins');
 const { userModel } = require('../../../models/index');
+const { AccountError } = require('../../errors');
 
 
 
 // ==============================
 //  CONTROLLERS
 // ==============================
-authentication.signup = {
+exports.account.signup = {
     get: (req, res) => {
         if (req.isAuthenticated()) return res.redirect('/home');
         return res.render('./home/account/signup');
@@ -28,17 +29,16 @@ authentication.signup = {
     }],
 };
 
-authentication.signin = {
+exports.account.signin = {
     get: (req, res) => {
         if (req.isAuthenticated()) return res.redirect('/home');
         return res.render('./home/account/signin');
     },
     post: (req, res, next) => {
         if (req.isAuthenticated()) return res.redirect('/home');
-        const passport = require('passport');
-        return passport.authenticate('local', (err, authUser) => {
+        return require('passport').authenticate('local', (err, authUser) => {
             if (err) return next(err);
-            if (!authUser) return next(new Error('Wrong email/username or password!'));
+            if (!authUser) return next(new AccountError('Wrong email/username or password!'));
             return req.logIn(authUser, err => {
                 if (err) return next(err);
                 req.session.cookie.expires = req.body.isPersisted ? new Date(Date.now() + 14 * 24 * 3600000) : false;
@@ -52,7 +52,7 @@ authentication.signin = {
     },
 };
 
-authentication.signout = {
+exports.account.signout = {
     get: (req, res) => {
         if (req.isAuthenticated()) {
             req.logout();