+++ routers/account.js
@@ -1,29 +1,29 @@
 const { RouterHub } = require('../controllers/routers/driver');
-const user = require('../controllers/routers/rules/user');
+const { account } = require('../controllers/routers/rules/account');
 
 
 
 // ==============================
 //  ROUTER HUB
 // ==============================
-const AuthenticationRouter = new RouterHub([{
+const UserRouter = new RouterHub([{
     route:          '/signup',
-    controller:     user.signup,
+    controller:     account.signup,
     settings:       { title: 'Sign Up' },
 }, {
     route:          '/signin',
-    controller:     user.signin,
+    controller:     account.signin,
     settings:       { title: 'Sign In' },
 }, {
     route:          '/signout',
-    controller:     user.signout,
+    controller:     account.signout,
 }]);
 
 
 // pre-used middleware
-AuthenticationRouter.use(require('../controllers/middleware/plugins')._md.usePassport);
+UserRouter.use(require('../controllers/middleware/plugins')._md.usePassport);
 
 
 
 // router exports
-module.exports = AuthenticationRouter.run();
+module.exports = UserRouter.run();